\name{getGDsig} 
\alias{getGDsig} 
\title{panConusig::getGDsig} 
\usage{getGDsig(sigs)}
\arguments{
\item{sigs}{Matrix. 48xn matrix of signature definitions, where n is the number of signatures}
}
\value{
48xn matrix of artificially genome doubled signature definitions.
} 
\description{
Artificially genome doubles a set of copy number signature definitions.
}
\details{
  
}

