\name{setupArms} 
\alias{setupArms} 
\title{panConusig::setupArms} 
\usage{setupArms(windowGR)}
\arguments{
\item{windowGR}{GenomicRanges object of genomic bins that the genome has been split into. See setupWindows().}
}
\value{
A list of chromosome arm information as a matrix ([[arms]]), chromosome arm information as a GenomicRanges object ([[armsGR]]) and overlaps between chromosome arms and genomic bins ([[armsOverlaps]]).
} 
\description{
Setup chromosome arm information for use in plotting recurrent signatures.
}
\details{
  
}

